Fixed Jack-related dumb-inception
The commit title is a joke on my previous commit.
Anyway, WoodTypeFactory is completely changed, for the most part. There was a glaring oversight with blocks and registering stuff, so I fixed it. In the process, the class is no longer a record, and the behavior is kind of the opposite of what it was before. It's still pretty intuitive, even more so, perhaps.